moved each component to its respective file
